Bug #3624: History cache should be stored in Java DB
Performance optimization: Make sure that the index cardinality
statistics in the database are up to date after inserting lots of
data. Without this change, we have seen that get() could be very slow
because the optimizer picks a bad plan based on stale statistics.

This adds a call to a system procedure introduced in Derby 10.5 (which
is still not released). On older versions of Derby we don't call the
procedure and we may therefore still see the performance problems.